Ballistic class refactoring
